+++ src/pvenv/subcommands/base.py
@@ -1,4 +1,7 @@
+import os
 from argparse import Namespace
+
+PREFIX = chr(2)
 
 
 class BaseCommand:
@@ -8,3 +11,9 @@ class BaseCommand:
 
     def run(self):
         raise NotImplementedError(f"{self.__class__.__name__} must implement run")
+
+    @staticmethod
+    def print(*args: str, **kwargs):  # noqa: A003
+        if not os.getenv("PVENV_DEBUG"):
+            args = tuple(f"{PREFIX}{arg}" for arg in args)
+        print(*args, **kwargs)